+++ lib/types.d.ts
@@ -1,0 +1,2 @@
+declare module 'vue3-virtual-scroller'
+declare module 'vue-grid-layout'